Added Focus Support
Enter/Leave are now triggered on FocustGained/FocusLost. This means that
in tmux when moving to a different pane all vim splits will be dimmed.
In order for this to work a terminal that supports focus events must be
used and the tmux option (focus-events on) must be set. For OSX users
iTerm2 supports focus events, in order to enable them the vim plugin
Vitality.vim is required as it sends escape codes to iTerm2 to turn on
focus reporting.